match

+++ src/main/java/com/fl/mapper/MatchMapper.java
@@ -21,7 +21,7 @@ public interface MatchMapper {
 	public List<MatchApplyDTO> listApplicant(Map<String, Object> map);
 	public MatchDTO findById(long num);
 	public List<TeamDTO> listUserTeams(long member_code);
-	public long getUserTeamRole(Map<String, Object> map);
+	public int getUserTeamRole(Map<String, Object> map);
 	public MatchDTO findByPrev(Map<String, Object> map);
 	public MatchDTO findByNext(Map<String, Object> map);
 	public void updateHitCount(long num) throws SQLException;
